+++ src/skill_tree/public/init.lua
@@ -1,0 +1,4 @@
+local require = require
+
+require 'skill_tree.public.test'
+require 'skill_tree.public.test1'