+++ src/skill_tree/random_node.lua
@@ -17,7 +17,7 @@ function Random:success()
     self:super().success(self)
 
     self._index_ = self._index_ + 1
-    if self._index_ <= #self._children then
+    if self._index_ <= #self._children_ then
         self:_run()
     else
         self:super():super().success(self)